requirements.txt
python-dotenv==1.0.1
pytest==8.3.5
